requirements.txt
fastapi==0.116.0
uvicorn[standard]==0.35.0
requests>=2.32
python-dotenv>=1.1
pandas>=2.2
numpy>=2.0
scipy>=1.13
scikit-learn>=1.8
xgboost>=3.0
mrmr-selection>=0.2.8
invoke>=2.2
